upgrade auth operation/ enhance the nav bar view and hide

+++ static/js/script.js
@@ -1,25 +1,19 @@
-const userName = document.getElementById("userName");
 const welcomeUser = document.getElementById("welcomeUser");
 const logoutBtn = document.getElementById("logout");
-const session= localStorage.getItem("userSession");
 
-//save the logged in user in local storage
-function get_user(){
-    const userSession = userName.value;
-    localStorage.setItem("userSession", userSession);
-}
 // remove the user from local storage on logout
-function remove_user(){
+logoutBtn.addEventListener("click", remove_user)
+if(window.location.href == "/"){
     localStorage.clear();
-}
-
-// show / hide the welcome user and log out from nav bar when the user logged in/out.
-if (session == null) {
     welcomeUser.style.display= "none";
     logoutBtn.style.display= "none";
 }
-else {
-    welcomeUser.innerHTML = "Hey " + session;
+
+function remove_user(){
+    localStorage.clear();
+    welcomeUser.style.display= "none";
+    logoutBtn.style.display= "none";
+    window.location.href = "/";
 }
 
 // show / hide message box "the message that's sent from the backend on any action of the user"
@@ -42,8 +36,9 @@ for (let i = 0; i < deleteBtns.length; i++) {
 }
 function confirmDelete(x){
     if (confirm ("Are you sure you want to delete this appointment")) {
-       window.location.href =`/delete?id=${x}`;
+       window.location.href =`/delete?id=${x}&userName=${userNameStorage}`;
     }
     else { 
     }
 }
+
